ENH: added cli option to load a catalog waveform data
This is useful if you want to use a catalog in a different machine where
the waveforms cannot be loaded.

Further enhancement, when the profiles are configured to never expire,
they are now loaded at startup. This makes the real time relocation
faster as only the new event data has to be fetched.